feat(amount): change amount type from int to long in order to increase maximum value permitted

+++ src/main/java/it/pagopa/transactions/commands/handlers/v2/TransactionRequestAuthorizationHandler.java
@@ -290,7 +290,7 @@ public class TransactionRequestAuthorizationHandler extends TransactionRequestAu
                                             t.getTransactionId().value(),
                                             new TransactionAuthorizationRequestData(
                                                     t.getPaymentNotices().stream()
-                                                            .mapToInt(
+                                                            .mapToLong(
                                                                     paymentNotice -> paymentNotice.transactionAmount()
                                                                             .value()
                                                             ).sum(),
